requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
pytest==6.2.4
PyJWT==1.7.1
PyYAML==5.4.1
SQLAlchemy==1.4.27
passlib[bcrypt]==1.7.4